Refac get google ad id
