+++ Adjust/adjust/src/main/java/com/adjust/sdk/Adjust.java
@@ -11,8 +11,6 @@ package com.adjust.sdk;
 
 import android.content.Context;
 import android.net.Uri;
-import android.os.AsyncTask;
-import android.os.Looper;
 
 /**
  * The main interface to Adjust.
@@ -78,38 +76,8 @@ public class Adjust {
         adjustInstance.setOfflineMode(enabled);
     }
 
-    public static void getGoogleAdId(Context context, final OnDeviceIdsRead onDeviceIdRead) {
-        ILogger logger = AdjustFactory.getLogger();
-        if (Looper.myLooper() != Looper.getMainLooper()) {
-            logger.debug("GoogleAdId being read in the background");
-            String GoogleAdId = Util.getPlayAdId(context);
-
-            logger.debug("GoogleAdId read " + GoogleAdId);
-            onDeviceIdRead.onGoogleAdIdRead(GoogleAdId);
-            return;
-        }
-        try{
-            logger.debug("GoogleAdId being read in the foreground");
-
-            new AsyncTask<Context,Void,String>() {
-                @Override
-                protected String doInBackground(Context... params) {
-                    ILogger logger = AdjustFactory.getLogger();
-                    Context innerContext = params[0];
-                    String innerResult = Util.getPlayAdId(innerContext);
-                    logger.debug("GoogleAdId read " + innerResult);
-                    return innerResult;
-                }
-
-                @Override
-                protected void onPostExecute(String playAdiId) {
-                    ILogger logger = AdjustFactory.getLogger();
-                    onDeviceIdRead.onGoogleAdIdRead(playAdiId);
-                }
-            }.execute(context);
-        } catch (Exception e) {
-            logger.error("Unable to get GoogleAdId from the foreground (%s)", e.getMessage());
-        }
+    public static void getGoogleAdId(Context context, OnDeviceIdsRead onDeviceIdRead) {
+        Util.getGoogleAdId(context, onDeviceIdRead);
     }
 }
